Moved the clustering to a package
