+++ deeplearning/src/main/java/nl/tudelft/bep/deeplearning/clustering/Mapping.java
@@ -1,4 +1,4 @@
-package nl.tudelft.bep.deeplearning;
+package nl.tudelft.bep.deeplearning.clustering;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
