+++ deeplearning/src/main/java/nl/tudelft/bep/deeplearning/clustering/NeighborDistance.java
@@ -1,4 +1,4 @@
-package nl.tudelft.bep.deeplearning;
+package nl.tudelft.bep.deeplearning.clustering;
 
 public class NeighborDistance {
